% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishoil.R
\docType{data}
\name{fishoil}
\alias{fishoil}
\title{Raman spectra of fish oil}
\format{A \code{data.frame} consisting of three parts. The estimated Iodine concentrations
of the fish oil samples, Raman spectra (\code{matrix}) and a replicate vector.}
\usage{
data(fishoil)
}
\description{
42 fish oil samples with 3 replicates each constitute a Raman data set of 126
samples over 3471 Raman shifts. There is a large fluorescence signal in the spectra that
needs to be removed before any data analysis can be performed.
}
\references{
K.H. Liland, A. Kohler, N.K. Afseth. Model-based pre-processing in Raman
 spectroscopy of biological samples. Journal of Raman Spectroscopy (2016).
}
